requirements.txt
pylint==2.17.4
bandit==1.7.5
radon==5.1.0
black==23.3.0
astunparse==1.6.3
flask==2.3.2
gitpython==3.1.43